# artifacts/support-pack/20260412T221950Z/openapi/auth-users-service.yaml
openapi: 3.0.3
info:
  title: SecureRAG Hub Auth Users Service API
  version: 1.0.0
  description: Persistent users, roles and permissions API for the Blade portal RBAC screens.
servers:
  - url: http://127.0.0.1:8091/api/v1
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
  /users:
    get:
      summary: List users
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
  /users/{user}:
    get:
      summary: Get user
      parameters:
        - $ref: "#/components/parameters/UserUuid"
      responses:
        "200": { description: User detail }
    put:
      summary: Update user
      parameters:
        - $ref: "#/components/parameters/UserUuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200": { description: User updated }
  /users/{user}/status:
    patch:
      summary: Update user status
      parameters:
        - $ref: "#/components/parameters/UserUuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [active, inactive, locked, pending_activation]
      responses:
        "200": { description: Status updated }
  /users/{user}/roles:
    post:
      summary: Attach roles to a user
      parameters:
        - $ref: "#/components/parameters/UserUuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roles]
              properties:
                roles:
                  type: array
                  items: { type: string }
      responses:
        "200": { description: Roles synchronized }
  /roles:
    get:
      summary: List roles
      responses:
        "200": { description: Roles list }
    post:
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "201": { description: Role created }
  /roles/{role}:
    get:
      summary: Get role
      parameters:
        - $ref: "#/components/parameters/RoleUuid"
      responses:
        "200": { description: Role detail }
    put:
      summary: Update role
      parameters:
        - $ref: "#/components/parameters/RoleUuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "200": { description: Role updated }
  /permissions:
    get:
      summary: List permissions
      responses:
        "200": { description: Permission catalogue }
components:
  parameters:
    UserUuid:
      in: path
      name: user
      required: true
      schema: { type: string, format: uuid }
    RoleUuid:
      in: path
      name: role
      required: true
      schema: { type: string, format: uuid }
  schemas:
    UserInput:
      type: object
      required: [first_name, last_name, email]
      properties:
        first_name: { type: string }
        last_name: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password }
        status: { type: string }
        department: { type: string }
        job_title: { type: string }
    RoleInput:
      type: object
      required: [name, label]
      properties:
        name: { type: string }
        label: { type: string }
        description: { type: string }
        permissions:
          type: array
          items: { type: string }
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items: { type: object }
        meta:
          type: object